% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bonferroni-Sidak_SCI.R
\name{BonferroniSCI}
\alias{BonferroniSCI}
\title{Bonferroni Correction for Simultaneous Confidence Interval}
\usage{
BonferroniSCI(
  theta,
  variances,
  alpha,
  n,
  sidak = FALSE,
  t_dist = FALSE,
  domains = NULL
)
}
\arguments{
\item{theta}{a vector of estimates of size Dx1.}

\item{variances}{a vector of variances for theta of size Dx1.}

\item{alpha}{a numeric value, 0 < alpha < 1, indicating the significance level.}

\item{n}{a numeric value for the overall sample size.}

\item{sidak}{a boolean value that indicates if the Sidak correction must be computed instead of the Bonferroni correction.}

\item{t_dist}{a boolean value that indicates if the t-distribution is used instead of the normal distribution for computing the quantiles.}

\item{domains}{an optional vectors of names for the domains of size Dx1; must be in the same order as theta.}
}
\description{
BonferroniSCI is a function to compute a simultaneous confidence interval (SCI) using the Bonferroni correction.
}
