% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{bootBooth}
\alias{bootBooth}
\title{Booth and al. (1994) bootstrap for finite population and simple random sampling}
\usage{
bootBooth(y, PI, N)
}
\description{
bootBooth is used to obtain a bootstrap replicate of a sample according to the method developped in Booth and al. (1994).
}
