% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{bootGross}
\alias{bootGross}
\title{Gross Bootstrap for finite population and simple random sampling}
\usage{
bootGross(y, PI, N)
}
\arguments{
\item{y}{a vector or a matrix of sampled data}

\item{PI}{a matrix of first and second order inclusion probabilities}

\item{N}{the size of the population}
}
\description{
bootGross is used to obtain a bootstrap replicate of a sample according to the method developped in Gross(1981).
}
