% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horvitz-Thompson.R
\name{HTestimator}
\alias{HTestimator}
\title{Horvitz-Thompson Estimator}
\usage{
HTestimator(y, weight, domains = NULL, mean = FALSE, Nd = NULL)
}
\arguments{
\item{y}{a vector of size nx1 (n is the sample size) of variables of interest.}

\item{weight}{a vector of size nx1 of sampling weights.}

\item{domains}{a vector of size nx1 of domains. If set to NULL, the function returns an estimator for the population.}

\item{mean}{a boolean variable that indicates if the function has to compute the mean instead of the total.}

\item{Nd}{a vector of size Dx1 (D is the number of domains) of domains population size for computing the means. If set to null, the function compute the Hajek estimators of the mean.}
}
\value{
A data frame with 3 columns : Domains, Estimates and Sample Sizes.
}
\description{
HTestimator is used to compute Horvitz-Thompson estimators of the total or of the mean for domains.
}
\examples{
y <- rnorm(100)
dom <- sample(1:5, size = 100, replace = TRUE)
Nd <- c(30,40,60,30,20)
weight <- sum(Nd) / 100
t_d <- HTestimator(y = y, weight = weight, domains = dom, mean = TRUE, Nd = Nd)
}
