% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horvitz-Thompson.R
\name{HTvar}
\alias{HTvar}
\title{Horvitz-Thompson Estimator of the Variance}
\usage{
HTvar(y, Weights, domains = NULL, mean = FALSE, Nd, method = 1)
}
\arguments{
\item{y}{a vector of size nx1 (n is the sample size) of variables of interest.}

\item{Weights}{a matrix of size nxn of sampling weights.}

\item{domains}{a vector of size nx1 of domains. If set to NULL, the function returns an estimator for the population.}

\item{mean}{a boolean variable that indicates if the function has to compute the mean instead of the total.}

\item{Nd}{a vector of size Dx1 (D is the number of domains) of domains population size for computing the means.}
}
\value{
A data frame with 3 columns : Domains, Variances and Sample Sizes.
}
\description{
HTvar is used to compute Horvitz-Thompson estimators of the variance for the total or for the mean for domains.
}
\examples{
n <- 100
y <- rnorm(n)
dom <- sample(1:5, size = n, replace = TRUE)
Nd <- c(30,40,60,30,20)
N <- sum(Nd)
pik <- n/N
pikl <- (n *(n-1)) / (N *(N-1))
Weights <- matrix(0, n, n)
diag(Weights) <- pik
Weights[upper.tri(Weights)] = PI[lower.tri(Weights)]  <- pikl
t_d_var <- HTvar(y = y, Weights = Weights, domains = dom, mean = TRUE, Nd = Nd)
}
