% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthetic_Data.R
\name{NERMData}
\alias{NERMData}
\title{Generation of synthethic data from a nested regression model}
\usage{
NERMData(N, dom, N_x, Dist_x, param_x, Beta, s_u, s_e, intercept = TRUE)
}
\arguments{
\item{N}{is the size of the population}

\item{dom}{is a N * 1 vector of domains that partioned the population}

\item{N_x}{is the number of auxilliary variables}

\item{Dist_x}{is the distributions of the auxilliary variables in the form r + distribution}

\item{param_x}{is a list of the parameters for each distribution in Dist_x}

\item{Beta}{is the N_x * 1 vector of beta's for the model}

\item{s_u}{is the standard error for the random effects}

\item{s_e}{is the standard error for the error terms}

\item{intercept}{is a boolean that indicates if the model has an intercept. By default it is set to TRUE.}
}
\description{
NERMData is a function to generate synthetic data from a nested error regression model.
}
